python_service/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/core_packages.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.115.8
    # via -r requirements/core_packages.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements/core_packages.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
jiter==0.8.2
    # via openai
openai==1.63.2
    # via -r requirements/core_packages.in
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pydantic==2.10.6
    # via
    #   -r requirements/core_packages.in
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements/core_packages.in
pytest==8.3.4
    # via
    #   -r requirements/core_packages.in
    #   pytest-asyncio
pytest-asyncio==0.25.3
    # via -r requirements/core_packages.in
python-dotenv==1.0.1
    # via
    #   -r requirements/core_packages.in
    #   pydantic-settings
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.45.3
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via -r requirements/core_packages.in
